LayoutTests/accessibility/loading-iframe-sends-notification-expected.txt

Before


After

End of test

This tests that when an iframe finishes loading, it sends a notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS findByAccessibleTitleSubstring(root, 'InnerButton') != null is false
Got notification on iframe.
PASS findByAccessibleTitleSubstring(root, 'InnerButton') != null is true
PASS gotIframeNotification is true
PASS successfullyParsed is true

TEST COMPLETE
